@peri044 2024-07-31
@HolyWu Is there any issue with aten.dot being lowered ? Generally, we intend to write converters for aten core ops https://pytorch.org/docs/main/torch.compiler_ir.html and the other ops are expected to be decomposed.