@HolyWu 2024-08-01
No issue for me. I just happened to find that the matrix layer in TRT has support for this operation and the code in `impl.matmul` already uses that operation. Then why not?