@peri044 2024-08-02
> No issue for me. I just happened to find that the matrix layer in TRT has support for this operation and the code in `impl.matmul` already uses that operation. Then why not?

I see. We prefer Pytorch lowering whenever it is feasible (especially for non aten core ops) since this would ensure our converter library to be light weighted. We can close this PR. If there are any performance concerns, let us know.