@HolyWu 2024-08-02
@peri044 One more question. If an operator was once in Core ATen IR but later was removed from Core ATen IR, for example `aten.roll` and `aten.pixel_shuffle`, then should the converters for those operators be removed so as to use PyTorch decomposition?